suppression Repo compte pour repo admin+ repo user

+++ rico-party-boot/src/main/java/rico/repo/IMessageRepo.java
@@ -9,10 +9,10 @@ import rico.model.Message;
 
 
 
-public interface IMessageRepo extends JpaRepository<Message, Long>{
+public interface IMessageRepo extends JpaRepository<Message, Integer>{
 	
 	@Query("Select m from Message m where m.event.id=?1")
-	List<Message> findAllByIdEvent(int idEvent);
+	List<Message> findAllByIdEvent(Integer idEvent);
 	
 
 }
